+++ modules/mbsimControl/mbsimControl/signal_sensors.cc
@@ -1,0 +1,51 @@
+/* Copyright (C) 2004-2025 MBSim Development Team
+ *
+ * This library is free software; you can redistribute it and/or 
+ * modify it under the terms of the GNU Lesser General Public 
+ * License as published by the Free Software Foundation; either 
+ * version 2.1 of the License, or (at your option) any later version. 
+ *  
+ * This library is distributed in the hope that it will be useful, 
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of 
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU 
+ * Lesser General Public License for more details. 
+ *  
+ * You should have received a copy of the GNU Lesser General Public 
+ * License along with this library; if not, write to the Free Software 
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301 USA
+ *
+ * Contact: markus.ms.schneider@gmail.com
+ */
+
+#include <config.h>
+#include "mbsimControl/signal_sensors.h"
+
+using namespace std;
+using namespace fmatvec;
+using namespace MBSim;
+using namespace MBXMLUtils;
+using namespace xercesc;
+
+namespace MBSimControl {
+
+  void SignalSensor::initializeUsingXML(DOMElement * element) {
+    Sensor::initializeUsingXML(element);
+    DOMElement *e=E(element)->getFirstElementChildNamed(MBSIMCONTROL%"signal");
+    saved_signal=E(e)->getAttribute("ref");
+  }
+
+  void SignalSensor::init(InitStage stage, const InitConfigSet &config) {
+    if(stage==resolveStringRef) {
+      if(not saved_signal.empty())
+        setSignal(getByPath<Signal>(saved_signal));
+      if(not signal)
+        throwError("Signal is not given!");
+    }
+    Sensor::init(stage, config);
+  }
+
+  int SignalSensor::getSignalSize() const {
+    return signal->getSignalSize();
+  }
+
+}